Add files via upload
Question: Write a function is_valid(s) that takes a string s as input and returns True if the string contains valid parentheses (i.e., (), [], {}) and False otherwise. A string is considered valid if every opening bracket has a corresponding closing bracket of the same type, and they are properly nested.

For example, the following strings are valid:

"{()[]}"
"{(){}[]}"
"([()])"
And the following strings are invalid:

"[]{(})"
"([)]"
Can you write a Python function to solve this problem?